Fix use of depreciated Hash class
We were using the depreciated Hash class in the move to destination
function of editing a transaction causing the transaction to work, but
the jump to another location to fail.

It also appears that we were using Hash in the queue class (queues
requests).  I have rewritten that class to use the link:'chain'
capability of request, which other programs (particularly airhock)
have proven work, and should be more efficient.